# MM-29981 - optimize markdown.Inspect