@agnivade 2020-11-04
This showed up in the flame graphs from load tests. This gets called in the createPost path, and within that, a lot of time was being spent in `markdown.Inspect`. The `regexp.MustCompile` was the main culprit. I just spent some more time optimizing more things I could find.