@lindalumitchell 2020-11-14
@agnivade sorry for the delay on this one! I ran some Markdown tests, and all looks good except for syntax rendering/highlighting. There's a test file we load using `/test url test-syntax-highlighting`, and below are screenshots of what it looks like on this PR's test server (can view the post here: https://mattermost-server-pr-16188.test.mattermost.cloud/pr16188/pl/dacfp99hkif9zgsqxr5z3ttyqy), then what the expected looks like on master.

This PR:
<img width="1203" alt="unexpected-missing-code-syntax-highlighting" src="https://user-images.githubusercontent.com/25826533/99132184-2936a280-25ca-11eb-9960-d5aac9d056b7.png">

Expected on master:
<img width="1342" alt="master-expected-code-syntax-highlighting" src="https://user-images.githubusercontent.com/25826533/99132291-8d596680-25ca-11eb-817a-8bf71cf26bb8.png">