@agnivade 2020-11-14
@lindalumitchell - That's a bit strange, because syntax highlighting is done on the client side. There is nothing on the server that does that. I just tried that post in my local setup and it works fine. Not really sure why it's not showing up in the PR.

Re: latex, you'd have to enable that in the system console. I just did that and it works.

In short, this PR does not modify the post at all. As long as the raw message is the same in both cases, then it is unlikely that server is involved here.

@hmhealey - Any ideas as to why syntax highlighting won't work?